+++ tests/test_scan.py
@@ -13,7 +13,7 @@ from NPTFit import psf_correction as pc # module for determining the PSF correct
 def test_scan_non_poiss():
     n = nptfit.NPTF(tag='Test_NPoiss')
 
-    fermi_data = np.load('fermi_data/fermidata_counts.npy').astype(int)
+    fermi_data = np.load('fermi_data/fermidata_counts.npy').astype(np.int32)
     fermi_exposure = np.load('fermi_data/fermidata_exposure.npy')
     n.load_data(fermi_data, fermi_exposure)
 
@@ -46,7 +46,7 @@ def test_scan_non_poiss():
 def test_scan_poiss():
     n = nptfit.NPTF(tag='Test_Poiss')
 
-    fermi_data = np.load('fermi_data/fermidata_counts.npy').astype(int)
+    fermi_data = np.load('fermi_data/fermidata_counts.npy').astype(np.int32)
     fermi_exposure = np.load('fermi_data/fermidata_exposure.npy')
     n.load_data(fermi_data, fermi_exposure)
